+++ Retina/Retina/Replace/Nodes/Reverse.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace Retina.Replace.Nodes
 {
@@ -19,7 +20,7 @@ namespace Retina.Replace.Nodes
             return new string(charArray);
         }
 
-        public override int GetLength(string input, List<MatchContext> matches, List<MatchContext> separators, int index)
+        public override BigInteger GetLength(string input, List<MatchContext> matches, List<MatchContext> separators, int index)
         {
             return Child.GetLength(input, matches, separators, index);
         }
