+++ Retina/Retina/Replace/Nodes/UpperCaseAll.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 
 namespace Retina.Replace.Nodes
 {
@@ -16,7 +17,7 @@ namespace Retina.Replace.Nodes
             return Child.GetString(input, matches, separators, index).ToUpperInvariant();
         }
 
-        public override int GetLength(string input, List<MatchContext> matches, List<MatchContext> separators, int index)
+        public override BigInteger GetLength(string input, List<MatchContext> matches, List<MatchContext> separators, int index)
         {
             return Child.GetLength(input, matches, separators, index);
         }
